requirements.txt
paho-mqtt
smbus2
chirp_modbus

